+++ lib/database/pokemon_database/pokemon_database_interraction.dart
@@ -11,7 +11,7 @@ class PokemonDatabase{
   }
 
   // create new pokemon
-  static Future<void> addPokemon(int idPoke, String name, int lifePoints, int pokedexId, int type, String imageURL, int size, int weight, int idAttack, int numberPokemon) async {
+  static Future<void> addPokemon(int idPoke, String name, double lifePoints, int pokedexId, int type, String imageURL, double size, double weight, int idAttack, int numberPokemon) async {
     final newPokemon = PokemonData()
       ..idPoke = idPoke
       ..name = name
